Consolidate DelegatingResourceDefinition
